requirements.txt
azure-functions==1.0.0b3
azure-functions-worker==1.0.0b4
beautifulsoup4==4.8.2
blis==0.4.1
Bottleneck==1.3.2
catalogue==1.0.0
certifi==2019.11.28
chardet==3.0.4
cycler==0.10.0
cymem==2.0.3
dataclasses==0.7
fastai==1.0.51
fastprogress==0.2.2
ffmpeg==1.4
ffmpeg-python==0.1.17
future==0.18.2
grpcio==1.14.2
grpcio-tools==1.14.2
idna==2.9
importlib-metadata==1.5.0
kiwisolver==1.1.0
matplotlib==3.2.0
murmurhash==1.0.2
numexpr==2.7.1
numpy==1.18.1
nvidia-ml-py3==7.352.0
opencv-python==4.2.0.32
packaging==20.3
pandas==1.0.1
Pillow==6.2.1
plac==1.1.3
preshed==3.0.2
protobuf==3.11.3
pyparsing==2.4.6
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3
requests==2.23.0
scipy==1.4.1
six==1.14.0
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
thinc==7.4.0
torch==1.0.1
torchvision==0.2.2
tqdm==4.19.9
typing==3.7.4.1
urllib3==1.25.8
wasabi==0.6.0
zipp==3.1.0
